Fix return type

+++ main.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 
 // x の平方根を返す関数。x >= 0 ではないとき exit(1) とする。
-int myRoot(double x) {
+double myRoot(double x) { // fix return type
     int y;
     if (x <= 0) {
         exit(1);
